The Big Merge TM
Changelog 20/04/2023

All of my created assets and level layouts have been transferred to / replicated on the dev project.

A lot of work has been done to the camera that follows the player

Added a few examples of how each layer will function

Added annotation

Everything is formatted nice and neatly.

All is functional for the next stage of the project. Assets will now be incredibly easy to import in and the devs can work simultaneously mechanics.