Adding in placeholder layer 5 detail
placeholder for now, i can work on different variants after getting layer 4 in place. Imported Leo's big pipe background in to engine, made a duplicate of the default sprite material and changed it so it could be affected by light, ready for lighting adjustments in the future. This will be helpful when producing consistent light throughout the level when using sprites. I think the big pipe background needs to have a depth of field effect so its blurred but the playspace layer still remains in focus.