further lighting tomfoolery
joe, when you have the time litter the scene with assets, not too much clutter but make it luved, feel free to change the light direction CTRL L to best suit your needs, some nice light effecs would be swell, eg light through corrugated metal and since we have lumen a slowly spinning fan? make sure the play are is well lit and the background isent too distracting. i tweaked the pp but feel free to change that too i was just messing with it
puddles would be cool.